requirements.txt
streamlit>=1.32.0
numpy>=1.26.0
scipy>=1.12.0
scikit-learn>=1.4.0
sentence-transformers>=2.7.0
torch>=2.2.0
transformers>=4.40.0
plotly>=5.20.0
pandas>=2.2.0
nltk>=3.8.1
python-docx>=1.1.0
pypdf>=4.0.0
tiktoken>=0.6.0
colorama>=0.4.6
tqdm>=4.66.0

# LLM providers (install the one you need)
openai>=1.30.0
anthropic>=0.25.0
groq>=0.11.0
